add config processing and command execution
This patch brings the program to its first fully working test version.
It now can read the default config file correctly, run the external
commands, concatenate their results to a string and finally display it
as a notification. The TODO list has been updated with the things to
work on next.